Load settings before rendering to avoid chart flickering on page load
When normal chart view is selected, the isometric view is rendered anyway. This allows to
instantly switch to it later, but if we load settings asynchonously, the chart flickers
on page load: the isometric view is switched on by default, and then turned off once
the settings are loaded.

This commit fixes that problem by moving settings loading to a separate method and
calling `renderIsometricChart` and `initUI` in its callback.